engine_traits: move WriteBatch to a module
