+++ components/engine_traits/src/engine.rs
@@ -4,17 +4,6 @@ use std::fmt::Debug;
 
 use crate::*;
 
-pub trait WriteBatch: Mutable + Send {
-    fn data_size(&self) -> usize;
-    fn count(&self) -> usize;
-    fn is_empty(&self) -> bool;
-    fn clear(&self);
-
-    fn set_save_point(&mut self);
-    fn pop_save_point(&mut self) -> Result<()>;
-    fn rollback_to_save_point(&mut self) -> Result<()>;
-}
-
 pub trait KvEngine:
     Peekable
     + Mutable
